+++ seq2science/util.py
@@ -6,6 +6,7 @@ import re
 import os
 import sys
 import glob
+import json
 import time
 import colorsys
 import tempfile
@@ -128,7 +129,7 @@ def samples2metadata_local(samples: List[str], config: dict, logger) -> dict:
         ):
             SAMPLEDICT[sample] = dict()
             SAMPLEDICT[sample]["layout"] = "PAIRED"
-        elif sample.startswith(("GSM", "DRX", "ERX", "SRX", "DRR", "ERR", "SRR", "CRX")):
+        elif sample.startswith(("GSM", "DRX", "ERX", "SRX", "DRR", "ERR", "SRR", "CRX", "ENCSR", "ENCFF")):
             continue
         else:
             extend_msg = ""
@@ -143,7 +144,7 @@ def samples2metadata_local(samples: List[str], config: dict, logger) -> dict:
                 f"We checked directory '{config['fastq_dir']}' "
                 f"for gzipped files starting with '{sample}' and containing '{config['fqsuffix']}'.\n"
                 + extend_msg
-                + f"Since the sample did not start with either GSM, SRX, SRR, ERR, and DRR we "
+                + f"Since the sample did not start with either GSM, SRX, SRR, ERX, ERR, DRX, DRR, CRX, ENCSR, or ENCFF we "
                 f"couldn't find it online..\n"
             )
             os._exit(1)  # noqa
@@ -207,7 +208,7 @@ def samples2metadata_gsa(samples: List[str], logger) -> dict:
                         "runs": ["CRR1234", "CRR4321"],
                         "gsa_fastq_http": {CRR1234: [link_1, link_2], ...,
 
-            "SRR5678": {"layout": "SINGLE",
+            "CRR5678": {"layout": "SINGLE",
                         "runs": ["CRR5678"],
                         gsa_fastq_http: {CRR5678: [link_1]},
             ...
@@ -362,6 +363,110 @@ def samples2metadata_sra(samples: List[str], logger) -> dict:
     return sampledict
 
 
+def samples2metadata_encode(samples: List[str], logger) -> dict:
+    """
+
+
+    """
+    metadata = {}
+    for sample in samples:
+        if sample.startswith("ENCSR"):
+            metadata.update(ENCSR2metadata(sample))
+        elif sample.startswith("ENCFF"):
+            metadata.update(ENCFF2metadata(sample))
+        else:
+            assert False
+    return metadata
+
+
+def ENCFF2metadata(sample):
+    """
+    encode file (fastq) to metadata
+    """
+    url = f"https://www.encodeproject.org/files/{sample}/?format=json"
+    response = urllib.request.urlopen(urllib.request.Request(url)).read()
+    response = json.loads(response.decode('utf-8'))
+    response
+
+    assert response["file_format"] == "fastq"
+
+    metadata = {sample: {}}
+
+    if response["run_type"] == "single-ended":
+        metadata[sample]["layout"] = "SINGLE"
+        metadata[sample]["encode_fastq_http"] = "https://www.encodeproject.org" + response["href"]
+        metadata[sample]["runs"] = [sample]
+    elif response["run_type"] == "paired-ended":
+        metadata[sample]["layout"] = "PAIRED"
+        mate = response["paired_with"].split("/")[2]
+        mate_url = f"https://www.encodeproject.org/files/{mate}/?format=json"
+        mate_response = urllib.request.urlopen(urllib.request.Request(mate_url)).read()
+        mate_response = json.loads(mate_response.decode('utf-8'))
+
+        if response["paired_end"] == "1":
+            metadata[sample]["encode_fastq_http"] = {
+                sample: [
+                    "https://www.encodeproject.org" + response["href"],
+                    "https://www.encodeproject.org" + mate_response["href"],
+                ]
+            }
+        else:
+            metadata[sample]["encode_fastq_http"] = {
+                sample: [
+                    "https://www.encodeproject.org" + mate_response["href"],
+                    "https://www.encodeproject.org" + response["href"],
+                ]
+            }
+
+        metadata[sample]["runs"] = [sample]
+    else:
+        assert False
+    return metadata
+
+
+def ENCSR2metadata(sample):
+    """
+    encode assay to metadata
+
+    an assay can exist out of multiple isogenic and technical replicates
+    """
+    url = f"https://www.encodeproject.org/search/?type=File&dataset=/experiments/{sample}/&file_format=fastq&format=json&frame=object"
+    response = urllib.request.urlopen(urllib.request.Request(url)).read()
+    response = json.loads(response.decode('utf-8'))
+    response
+
+    metadata = {sample: {}}
+
+    # check if all run types are the same
+    assert len({file["run_type"] for file in response["@graph"]}) == 1
+    layout = response["@graph"][0]["run_type"]
+
+    if layout == "single-ended":
+        metadata[sample]["layout"] = "SINGLE"
+        metadata[sample]["encode_fastq_http"] = {
+            x["accession"]: "https://www.encodeproject.org" + x["href"] for x in response["@graph"]
+        }
+        metadata[sample]["runs"] = list(metadata[sample]["encode_fastq_http"].keys())
+
+    elif layout == "paired-ended":
+        graph = {x["accession"]: x for x in response["@graph"]}
+
+        metadata[sample]["layout"] = "PAIRED"
+        metadata[sample]["runs"] = [x["accession"] for x in response["@graph"] if x["paired_end"] == "1"]
+
+        metadata[sample]["encode_fastq_http"] = {
+            run: [
+                "https://www.encodeproject.org" + graph[run]["href"],
+                "https://www.encodeproject.org" + graph[graph[run]["paired_with"].split("/")[2]]["href"],
+            ]
+            for run in metadata[sample]["runs"]
+        }
+    else:
+        assert False
+
+    return metadata
+
+
 def samples2metadata(samples: List[str], config: dict, logger) -> dict:
     local_samples = samples2metadata_local(samples, config, logger)
     public_samples = [sample for sample in samples if sample not in local_samples.keys()]
@@ -371,7 +476,9 @@ def samples2metadata(samples: List[str], config: dict, logger) -> dict:
 
     gsa_samples = [sample for sample in public_samples if sample.startswith("CRX")]
     gsa_samples = samples2metadata_gsa(gsa_samples, logger)
-    rest_public_samples = [sample for sample in public_samples if sample not in gsa_samples]
+    encode_samples = [sample for sample in public_samples if sample.startswith(("ENCSR", "ENCFF"))]
+    encode_samples = samples2metadata_encode(encode_samples, logger)
+    rest_public_samples = [sample for sample in public_samples if sample not in (list(gsa_samples.keys()) + list(encode_samples.keys()))]
 
     # chop public samples into smaller chunks, doing large queries results into
     # pysradb decode errors..
@@ -383,7 +490,7 @@ def samples2metadata(samples: List[str], config: dict, logger) -> dict:
         # just to be sure sleep in between to not go over our API limit
         time.sleep(1)
 
-    return {**local_samples, **sra_samples, **gsa_samples}
+    return {**local_samples, **sra_samples, **gsa_samples, **encode_samples}
 
 
 def sieve_bam(configdict):